Contributed by applekor     * Clarification: Two or more tokens with a common macro now add that macro to the common macro groups.     * Bug Fix: Common macros in the Selection panel now function as they did in b41. If you execute a common macro, it looks for the macro of the same name on each selected token and only executes those matching individual macros against their respective tokens. Tokens without a matching macro are not executed against. If any matching individual macro has "Apply to Selected Tokens" checked, it should still execute against ALL selected tokens.     * New Feature: An "Options" tab has been added to the macro editor dialog. The macro details now appear on a "Details" tab.     * New Feature: The criteria for comparing macro details when determining similarity for populating common macros on the Selection panel and for drag-and-drop operations are no longer strictly hard coded. The user can choose to have individual macros ignore Group, Sort Prefix, Command, Include Label, Auto Execute, and/or Apply to Selected Tokens when comparing against other macros by clearing respective check boxes on the new "Options" tab of the macro's editor dialog.     * New Feature: GM's can designate any individual macro as player editable/non-editable by setting/clearing a new check box in the new "Options" tab of the macro's editor dialog. Certain popup menu items are also made unavailable to players when this check box is set.     * New/Returning Feature: Common macro buttons/groups now have popup menus on right-clicks. Through this interface change, macros can be added/removed to all selected tokens simultaneously. Additionally, macro import/export features have been extended to common macros.     * New Feature: When hovering over a macro button, tokens that would be affected by execution of the macro are now highlighted with a new, thicker, red & green selection border. This feature isn't fully implemented yet for the Impersonate panel. The Global, Campaign, and Selection panels appear to have full functionality.

+++ src/net/rptools/maptool/client/ui/macrobuttons/buttons/MacroButtonPopupMenu.java
@@ -23,10 +23,14 @@ import java.awt.EventQueue;
 import java.awt.event.ActionEvent;
 import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 
+import net.rptools.maptool.client.AppUtil;
 import net.rptools.maptool.client.MapTool;
 import net.rptools.maptool.client.ui.MacroButtonDialog;
 import net.rptools.maptool.model.MacroButtonProperties;
+import net.rptools.maptool.model.Token;
 import net.rptools.maptool.util.PersistenceUtil;
 
 @SuppressWarnings("serial")
@@ -35,25 +39,66 @@ public class MacroButtonPopupMenu extends JPopupMenu{
 	private final MacroButton button;
 	private final String panelClass;
 	
-	public MacroButtonPopupMenu(MacroButton parent, String panelClass) {
+	public MacroButtonPopupMenu(MacroButton parent, String panelClass, Boolean commonMacro) {
 		this.button = parent;
 		this.panelClass = panelClass;
-		add(new EditButtonAction());
-		add(new JSeparator());
-		add(new AddNewButtonAction());
-		add(new DuplicateButtonAction());
-		add(new JSeparator());
-		add(new ResetButtonAction());
-		add(new DeleteButtonAction());
-		add(new JSeparator());
-		add(new ExportMacroAction());
-		add(new JSeparator());
-		add(new RunMacroForEachSelectedTokenAction());
+		if(panelClass.equals("SelectionPanel")) {
+			if(MapTool.getFrame().getSelectionPanel().getCommonMacros().contains(button.getProperties())) {
+				addCommonActions();
+			} else {
+				addActions();
+			}
+		} else {
+			addActions();
+		}
+	}
+	
+	private void addActions() {
+		if(MapTool.getPlayer().isGM() || (!MapTool.getPlayer().isGM() && button.getProperties().getAllowPlayerEdits())) {
+			add(new EditButtonAction());
+			add(new JSeparator());
+			add(new AddNewButtonAction());
+			add(new DuplicateButtonAction());
+			add(new JSeparator());
+			add(new ResetButtonAction());
+			add(new DeleteButtonAction());
+			add(new JSeparator());
+			add(new ExportMacroAction());
+			add(new JSeparator());
+			add(new RunMacroForEachSelectedTokenAction());
+		} else {
+			add(new AddNewButtonAction());
+			add(new JSeparator());
+			add(new RunMacroForEachSelectedTokenAction());
+		}
+		
+	}
+	
+	private void addCommonActions() {
+		if(MapTool.getPlayer().isGM() || (!MapTool.getPlayer().isGM() && button.getProperties().getAllowPlayerEdits())) {
+			add(new EditButtonAction());
+			add(new AddNewButtonAction("Add New to Selected"));
+			add(new DuplicateButtonAction("Duplicate Macro on Selected"));
+			add(new JSeparator());
+			add(new DeleteButtonAction("Delete from Common"));
+			add(new JSeparator());
+			add(new ExportMacroAction("Export Common Macro"));
+			add(new JSeparator());
+			add(new RunMacroForEachSelectedTokenAction());
+		} else {
+			add(new AddNewButtonAction("Add New to Selected"));
+			add(new JSeparator());
+			add(new RunMacroForEachSelectedTokenAction());
+		}		
 	}
 
 	private class AddNewButtonAction extends AbstractAction {
 		public AddNewButtonAction() {
-			putValue(Action.NAME, "New");
+			putValue(Action.NAME, "Add New Macro");
+		}
+		
+		public AddNewButtonAction(String name) {
+			putValue(Action.NAME, name);
 		}
 
 		public void actionPerformed(ActionEvent event) {
@@ -62,6 +107,14 @@ public class MacroButtonPopupMenu extends JPopupMenu{
 				new MacroButtonProperties(panelClass, MacroButtonPrefs.getNextIndex(), button.getProperties().getGroup());
 			} else if (panelClass.equals("CampaignPanel")) {
 				new MacroButtonProperties(panelClass, MapTool.getCampaign().getMacroButtonNextIndex(), button.getProperties().getGroup());
+			} else if(panelClass.equals("SelectionPanel")) {
+				if(MapTool.getFrame().getSelectionPanel().getCommonMacros().contains(button.getProperties())) {
+					for(Token nextToken : MapTool.getFrame().getCurrentZoneRenderer().getSelectedTokensList()) {
+						new MacroButtonProperties(nextToken, nextToken.getMacroNextIndex(), button.getProperties().getGroup());
+					}
+				} else {
+					new MacroButtonProperties(button.getToken(), button.getToken().getMacroNextIndex(), button.getProperties().getGroup());
+				}
 			} else if (button.getToken()!= null){
 				new MacroButtonProperties(button.getToken(), button.getToken().getMacroNextIndex(), button.getProperties().getGroup());
 			}
@@ -74,17 +127,17 @@ public class MacroButtonPopupMenu extends JPopupMenu{
 		}
 
 		public void actionPerformed(ActionEvent event) {
-			if (button.getToken()== null){
-				new MacroButtonDialog(false).show(button);
-			} else {
-				new MacroButtonDialog(true).show(button);
-			}
+			new MacroButtonDialog().show(button);
 		}
 	}
 
 	private class DeleteButtonAction extends AbstractAction {
 		public DeleteButtonAction() {
 			putValue(Action.NAME, "Delete");
+		}
+		
+		public DeleteButtonAction(String name) {
+			putValue(Action.NAME, name);
 		}
 
 		public void actionPerformed(ActionEvent event) {
@@ -96,8 +149,31 @@ public class MacroButtonPopupMenu extends JPopupMenu{
 				MacroButtonPrefs.delete(button.getProperties());
 			} else if (panelClass.equals("CampaignPanel")) {
 				MapTool.getCampaign().deleteMacroButton(button.getProperties());
+			} else if(panelClass.equals("SelectionPanel")) {
+				if(MapTool.getFrame().getSelectionPanel().getCommonMacros().contains(button.getProperties())) {
+					for(Token nextToken : MapTool.getFrame().getCurrentZoneRenderer().getSelectedTokensList()) {
+						if(AppUtil.playerOwns(nextToken)) {
+							List<MacroButtonProperties> workingMacros = new ArrayList<MacroButtonProperties>();
+							Boolean hashCodesMatch = false;
+							Boolean allowDelete = false;
+							for(MacroButtonProperties nextMacro : nextToken.getMacroList(true)) {
+								hashCodesMatch = nextMacro.hashCodeForComparison() == button.getProperties().hashCodeForComparison();
+								allowDelete = MapTool.getPlayer().isGM() || (!MapTool.getPlayer().isGM() && nextMacro.getAllowPlayerEdits());
+								if(!hashCodesMatch || !allowDelete) {
+									workingMacros.add(nextMacro);
+								}
+							}
+							nextToken.replaceMacroList(workingMacros);
+						}
+					}
+				} else {
+					button.getToken().deleteMacroButtonProperty(button.getProperties());
+				}
+				MapTool.getFrame().getSelectionPanel().reset();
 			} else if (button.getToken()!= null){
-				button.getToken().deleteMacroButtonProperty(button.getProperties());
+				if(AppUtil.playerOwns(button.getToken())) {
+					button.getToken().deleteMacroButtonProperty(button.getProperties());
+				}
 			}
 		}
 	}
@@ -106,12 +182,24 @@ public class MacroButtonPopupMenu extends JPopupMenu{
 		public DuplicateButtonAction() {
 			putValue(Action.NAME, "Duplicate");
 		}
+		
+		public DuplicateButtonAction(String name) {
+			putValue(Action.NAME, name);
+		}
 
 		public void actionPerformed(ActionEvent event) {
 			if (panelClass.equals("GlobalPanel")) {
 				new MacroButtonProperties(panelClass, MacroButtonPrefs.getNextIndex(), button.getProperties());
 			} else if (panelClass.equals("CampaignPanel")) {
 				new MacroButtonProperties(panelClass, MapTool.getCampaign().getMacroButtonNextIndex(), button.getProperties());
+			} else if(panelClass.equals("SelectionPanel")) {
+				if(MapTool.getFrame().getSelectionPanel().getCommonMacros().contains(button.getProperties())) {
+					for(Token nextToken : MapTool.getFrame().getCurrentZoneRenderer().getSelectedTokensList()) {
+						new MacroButtonProperties(nextToken, nextToken.getMacroNextIndex(), button.getProperties());
+					}
+				} else {
+					new MacroButtonProperties(button.getToken(), button.getToken().getMacroNextIndex(), button.getProperties());
+				}
 			} else if (button.getToken() != null){
 				new MacroButtonProperties(button.getToken(), button.getToken().getMacroNextIndex(), button.getProperties());
 			}
@@ -144,6 +232,10 @@ public class MacroButtonPopupMenu extends JPopupMenu{
 				putValue(Action.NAME, "Export Macro");
 			}
 			
+			private ExportMacroAction(String name) {
+				putValue(Action.NAME, name);
+			}
+			
 			public void actionPerformed(ActionEvent event) {
 				JFileChooser chooser = MapTool.getFrame().getSaveMacroFileChooser();
 				
@@ -165,6 +257,18 @@ public class MacroButtonPopupMenu extends JPopupMenu{
 						}
 						
 						try {
+							if(panelClass.equals("SelectionPanel")) {
+								if(MapTool.getFrame().getSelectionPanel().getCommonMacros().contains(button.getProperties())) {
+									if(confirmCommonExport(button.getProperties())) {
+										PersistenceUtil.saveMacro(button.getProperties(), selectedFile);										
+									} else {
+										MapTool.showInformation("Macro export cancelled.");
+										return;
+									}
+								} else {
+									PersistenceUtil.saveMacro(button.getProperties(), selectedFile);	
+								}
+							}
 							PersistenceUtil.saveMacro(button.getProperties(), selectedFile);				
 							MapTool.showInformation("Macro Saved.");
 						} catch (IOException ioe) {
@@ -174,5 +278,42 @@ public class MacroButtonPopupMenu extends JPopupMenu{
 					}
 				});				
 			}
+	}
+	
+	private Boolean confirmCommonExport(MacroButtonProperties buttonMacro) {
+		Boolean failComparison = false;
+		String comparisonResults = "";
+		if(!buttonMacro.getCompareGroup()) {
+			failComparison = true;
+			comparisonResults = comparisonResults + "<li>Group</li>";
+		}
+		if(!buttonMacro.getCompareSortPrefix()) {
+			failComparison = true;
+			comparisonResults = comparisonResults + "<li>Sort Prefix</li>";
+		}
+		if(!buttonMacro.getCompareCommand()) {
+			failComparison = true;
+			comparisonResults = comparisonResults + "<li>Command</li>";
+		}
+		if(!buttonMacro.getCompareIncludeLabel()) {
+			failComparison = true;
+			comparisonResults = comparisonResults + "<li>Include Label</li>";
+		}
+		if(!buttonMacro.getCompareAutoExecute()) {
+			failComparison = true;
+			comparisonResults = comparisonResults + "<li>Auto Execute</li>";
+		}
+		if(!buttonMacro.getApplyToTokens()) {
+			failComparison = true;
+			comparisonResults = comparisonResults + "<li>Apply to Selected Tokens</li>";
+		}
+		if(failComparison) {
+		failComparison = MapTool.confirm("<html><body>Macro \"" + buttonMacro.getLabel() +
+				"\" has the following comparison criteria deselected: <br><ul>" +
+				comparisonResults + "</ul><br>Do you wish to add this macro to " +
+				"the export list?<br><br>Selecting \"Yes\" will result in the " +
+				"macro being exported without the information listed.</body></html>");
+		}
+		return failComparison;
 	}
 }
